Coding Challenge
Added :
Create a webservice endpoint
a) To return a list of pets given an ownerId
b) Visit currently is not associated with a vet. Add an association between visit and vet (a visit maps to a vet). Create a new webservice endpoint that returns all pets that has visited a vet